requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Protein Structure
biopython>=1.81
biotite>=0.37.0

# Visualization
py3Dmol>=2.0.0
plotly>=5.14.0

# Notebook
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0

# ML/DL utilities
einops>=0.6.0
timm>=0.9.0

# Structure prediction (optional)
# dm-haiku>=0.0.10
# jax>=0.4.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0
